third_party/WebKit/LayoutTests/shadow-dom/query-selector.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/shadow-dom.js"></script>
<div id="x"><span>test1</span></div>
<div id="x"><span>test2</span></div>
<div id="host">
  <template data-mode="open">
    <div id="y"><span></span></div>
    <div id="y"><span></span></div>
    <div><span id="x" class="y">Can you find this?</span></div>
  </template>
</div>
<script>
convertTemplatesToShadowRootsWithin(host);
test(() => {
  assert_equals(document.querySelectorAll('#x').length, 2);
  assert_equals(document.querySelectorAll('#x span').length, 2);
}, 'querySelectorAll for multiple #Ids in a document tree');

test(() => {
  assert_equals(host.shadowRoot.querySelectorAll('#y').length, 2);
  assert_equals(host.shadowRoot.querySelectorAll('#y span').length, 2);
}, 'querySelectorAll for multiple #Ids in a shadow tree');

test(() => {
  var spans = document.querySelectorAll('body >>> #x');
  assert_equals(spans.length, 3);
  assert_equals(spans[0].textContent, 'test1');
  assert_equals(spans[1].textContent, 'test2');
  assert_equals(spans[2].textContent, 'Can you find this?');
}, 'deep combinator >>> should work for querySelectorAll.');

test(() => {
  var span = document.querySelector('body >>> .y');
  assert_equals(span.textContent, 'Can you find this?');
}, 'deep combinator >>> should work for querySelector');
</script>
